requirements.txt
face_recognition
pillow
image_hash
tkinter
